rename inproperly named package
